backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
opencv-python>=4.10.0.84
numpy>=1.26.2
Pillow>=10.3.0
python-multipart>=0.0.9
websockets>=12.0
torch>=2.0.0
